Separated POR implementations
- removed inferior SCC algorithm
