+++ src/pins-lib/Makefile.am
@@ -71,6 +71,13 @@ libpins_la_SOURCES += pins2pins-mucalc.c pins2pins-mucalc.h
 libpins_la_SOURCES += pins2pins-ltl.c
 libpins_la_SOURCES += pins2pins-por.h
 libpins_la_SOURCES += pins2pins-por.c
+libpins_la_SOURCES += por-ample.h
+libpins_la_SOURCES += por-ample.c
+libpins_la_SOURCES += por-beam.h
+libpins_la_SOURCES += por-beam.c
+libpins_la_SOURCES += por-deletion.h
+libpins_la_SOURCES += por-deletion.c
+libpins_la_SOURCES += por-internal.h
 libpins_la_SOURCES += pins2pins-por-check.c
 libpins_la_SOURCES += property-semantics.c property-semantics.h
 libpins_la_CPPFLAGS = -DNXT $(AM_CPPFLAGS) -I${top_srcdir}/ltl2ba
