+++ src/pins-lib/por-ample.h
@@ -1,0 +1,17 @@
+#ifndef AMPLE_POR
+#define AMPLE_POR
+
+
+#include <pins-lib/pins2pins-por.h>
+
+
+typedef struct ample_ctx_s ample_ctx_t;
+
+extern ample_ctx_t *ample_create_context (por_context *ctx, bool all);
+
+extern bool         ample_is_stubborn (por_context *ctx, int group);
+
+extern int          ample_search_all (model_t self, int *src, TransitionCB cb,
+                                      void *user_context);
+
+#endif
